// types.ts
export interface Supplier {
  id: string;
  _id?: string;
  supplierName: string;
  contactNumber: string;
  township?: string;
  address?: string;
  isDeleted?: boolean;
  deletedAt?: string | null;
  createdAt?: string;
  updatedAt?: string;
}